editing

+++ sort.h
@@ -37,7 +37,7 @@ void merge(int *array, int *left, int *right, size_t size);
 void heap_sort(int *array, size_t size);
 void exchange(int *x, int *y);
 void sort(int *a, int b);
-void heapify(int *a, int b, int c);
+void heapify(int *a, size_t b, size_t c);
